feat: 지역, 태그 데이터를 constants로 이동, SelectRegion 컴포넌트 스타일링
