--- src/components/common/SelectRegion/types.ts
@@ -1,3 +1,0 @@
-export type Region = {
-  text: string;
-};
