+++ src/components/common/index.ts
@@ -12,3 +12,4 @@ export { default as Field } from './Field';
 export { default as ErrorMessage } from './Form/ErrorMessage';
 export { default as SearchInput } from './SearchInput';
 export { default as SelectRegion } from './SelectRegion';
+export { default as SelectTags } from './SelectTags';
